refactor(modelcat): split catalog into table, ids, query, ladder
